UniversalRing: Merge abstract and concrete types
